We need framebuffers
Sadly, the transformations on the box had to go :(

Too bad, I was proud of having a rotating volume box.

This means the box geometry no longer scales with the box being calculated by the shader. We need access to the full screen to render the volume without doing tedious updates to the geometry and/or some sort of model matrix on the cube geometry.